rename words -> word
