+++ entities/word.py
